fixes from github comments

+++ app/models/tweet.rb
@@ -4,12 +4,16 @@ class Tweet < ActiveRecord::Base
   validates :user_id, presence: true
   validates :text, presence: true
 
+  has_many :user_like_tweets
+
   def liked_by(user)
-    UserLikeTweet.create(user_id: user.id, tweet_id: self.id);
+    user_like_tweets.create(user_id: user.id) unless user_like_tweets.exist? user_id: user_id
   end
 
   def unliked_by(user)
     like = UserLikeTweet.where(user_id: user.id, tweet_id: self.id)
-    like.destroy
+    if like.present?
+      like.destroy
+    end
   end
 end
